+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/request/VerifyReturnSupplyReqVo.java
@@ -9,6 +9,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.constraints.NotEmpty;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -90,14 +91,13 @@ public class VerifyReturnSupplyReqVo implements ILockStockReqVO {
     private Long preMainUnitNum;
 
     @ApiModelProperty("预计含税总金额")
-    private Long preTaxAmount;
+    private BigDecimal preTaxAmount;
 
     @ApiModelProperty("预计无税总金额")
-    private Long preAmount;
+    private BigDecimal preAmount;
 
     @ApiModelProperty("操作人")
     private String operator;
-
 
     @ApiModelProperty("商品详情vo")
     private List<ReturnSupplyItemReqVo> itemReqVos;
